requirements.txt
arrow==1.3.0
binaryornot==0.4.4
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cookiecutter==2.6.0
cycler==0.12.1
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-cytoscape==1.0.0
dash-extensions==1.0.14
dash-html-components==2.0.0
dash-svg==0.0.12
dash-table==5.0.0
dataclass-wizard==0.22.3
distlib==0.3.8
EditorConfig==0.12.4
filelock==3.13.3
Flask==3.0.2
Flask-Caching==2.2.0
fonttools==4.49.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsbeautifier==1.15.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==9.1.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
pyarrow==15.0.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
rich==13.7.1
s-gd2==1.8.1
scipy==1.12.0
shapely==2.0.4
six==1.16.0
tenacity==8.2.3
text-unidecode==1.3
tqdm==4.66.4
tulip-python==5.7.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.1
zipp==3.17.0
